/db: Track assignment variables to and from cluster
Various updates to database implementation and assignment interaction with said database.

---

Assigments can now be added to the MongoDB module collection.
A given assignment can be removed from a module collection.
db_helper class introduced with functionality to communicate with Assignment cluster on MongoDb.
assignment_parser function introduced to deal with BSON objects that are then translated into usable JSON objects.
Introduce "name" variable that is used to track assignment documents.
This name variable is used to track if an assignment with the same name is already in the collection
as to therefore bypass the possibility of assignment duplication.
test.py script updated to test these new "db_helper" functionality.

Signed-off-by: Brian <BrianArchbold96@yahoo.ie>