/db: add functionality for assignment_update functions
Added functionality for the rest of the assignment_update functions which were
previously commented out.
NOTE: Implementation is very similar for these functions, possible to refactor them
in future, a note has been made within the db_helper file and is a part of the previous commit
Also added tests for these Assignment_update functions in the test.py file.

Signed-off-by: Brian <BrianArchbold96@yahoo.ie>