mycroft-assignment: list assignments for given module
A user can now list assignments for a given module. This is especially useful if a user is
trying to update data for an assignment eg. "accumulated mark" and is unsure of the name given
to the module.
If there is no assignments for a module or a module_id is invalid, mycroft will say an appropriate
dialog

Signed-off-by: Brian <BrianArchbold96@yahoo.ie>